moving operations files to specific package
