+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/ReservesScreen.kt
@@ -65,7 +65,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
-import br.com.rbrthmn.ui.financialcompanion.components.AddOperationDialog
+import br.com.rbrthmn.ui.financialcompanion.screens.operations.components.AddOperationDialog
 import br.com.rbrthmn.ui.financialcompanion.navigation.NavigationDestination
 import br.com.rbrthmn.ui.financialcompanion.utils.OperationType
 import br.com.rbrthmn.ui.financialcompanion.utils.valueWithCurrencyString
