+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/operations/components/OperationsListCard.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.components
+package br.com.rbrthmn.ui.financialcompanion.screens.operations.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -64,7 +64,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
-import br.com.rbrthmn.ui.financialcompanion.screens.Operation
+import br.com.rbrthmn.ui.financialcompanion.components.DatePickerDocked
+import br.com.rbrthmn.ui.financialcompanion.components.ReservesDropdownMenu
+import br.com.rbrthmn.ui.financialcompanion.screens.operations.Operation
 import br.com.rbrthmn.ui.financialcompanion.utils.OperationType
 import br.com.rbrthmn.ui.financialcompanion.utils.getOperationsMock
 import java.text.SimpleDateFormat
